Refactor and remove ModernCalibrationWizard and related components
- Deleted the ModernCalibrationWizard class and its implementation files to streamline the GUI components.
- Removed associated UI elements such as ModernStepIndicator, ModernProgressRing, and ModernCameraPreview.
- Updated the EnhancedImageWidget to include a new ModernProgressIndicator class for progress visualization.
- Added new methods in EpipolarChecker for parameter change handling and visualization updates.
- Adjusted the main window to include necessary headers for the new structure.
- Enhanced edge case testing framework with concurrent access testing capabilities.
- Updated neural matcher edge case tests to reflect changes in the EnhancedNeuralMatcher API.
- Improved system edge case tests for better error logging and race condition detection.

+++ tests/test_system_edge_cases.cpp
@@ -316,7 +316,7 @@ TEST_F(SystemLevelEdgeCaseTest, RaceConditionDetection) {
     // Log some error messages for debugging
     if (!error_messages.empty()) {
         std::cout << "Sample concurrent operation errors:\n";
-        for (size_t i = 0; i < std::min(5u, error_messages.size()); ++i) {
+        for (size_t i = 0; i < std::min(static_cast<size_t>(5u), error_messages.size()); ++i) {
             std::cout << "  " << error_messages[i] << "\n";
         }
     }
